% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idler.R
\name{idle_timeout}
\alias{idle_timeout}
\title{Set up observer for client-side idle timeout message}
\usage{
idle_timeout(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{Session to observe (the default should almost always be used).}
}
\description{
Set up observer for client-side idle timeout message
}
