% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idler.R
\name{idle_warning}
\alias{idle_warning}
\title{Set up observer to warn client their session will be closed}
\usage{
idle_warning(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{Session to observe (the default should almost always be used)}
}
\description{
Set up observer to warn client their session will be closed
}
