% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idler.R
\name{set}
\alias{set}
\title{Set the idle timer duration in seconds}
\usage{
set(duration)
}
\arguments{
\item{duration}{number of seconds of inactivity before the session is ended}
}
\description{
Set the idle timer duration in seconds
}
