% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idler.R
\name{use_idler}
\alias{use_idler}
\title{Set up a Shiny app to use idler}
\usage{
use_idler()
}
\value{
shiny.tag containing a JS message handler to monitor client app use
}
\description{
This function must be called from a Shiny app's UI
}
\examples{
htmltools::tags$head(use_idler())
}
